+++ src/main/kotlin/org/wfanet/measurement/integration/common/reporting/v2/InProcessLifeOfAReportIntegrationTest.kt
@@ -62,12 +62,14 @@ import org.wfanet.measurement.api.v2alpha.Measurement
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumerKey
 import org.wfanet.measurement.api.v2alpha.MeasurementConsumersGrpcKt.MeasurementConsumersCoroutineStub
 import org.wfanet.measurement.api.v2alpha.MeasurementKt
+import org.wfanet.measurement.api.v2alpha.MeasurementsGrpcKt.MeasurementsCoroutineStub
 import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt
 import org.wfanet.measurement.api.v2alpha.batchGetEventGroupMetadataDescriptorsRequest
 import org.wfanet.measurement.api.v2alpha.eventGroup as cmmsEventGroup
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.Person
 import org.wfanet.measurement.api.v2alpha.getDataProviderRequest
 import org.wfanet.measurement.api.v2alpha.getMeasurementConsumerRequest
+import org.wfanet.measurement.api.v2alpha.listMeasurementsRequest
 import org.wfanet.measurement.api.v2alpha.testing.MeasurementResultSubject.Companion.assertThat
 import org.wfanet.measurement.api.withAuthenticationKey
 import org.wfanet.measurement.common.OpenEndTimeRange
@@ -388,6 +390,23 @@ abstract class InProcessLifeOfAReportIntegrationTest(
     ImpressionQualificationFiltersCoroutineStub(reportingServer.publicApiChannel)
   }
 
+  private val publicMeasurementsClient by lazy {
+    MeasurementsCoroutineStub(inProcessCmmsComponents.kingdom.publicApiChannel)
+  }
+
+  suspend fun listMeasurements(): List<Measurement> {
+    val measurementConsumerData = inProcessCmmsComponents.getMeasurementConsumerData()
+
+    val measurements: List<Measurement> = runBlocking {
+      publicMeasurementsClient
+        .withAuthenticationKey(measurementConsumerData.apiAuthenticationKey)
+        .listMeasurements(listMeasurementsRequest { parent = measurementConsumerData.name })
+        .measurementsList
+    }
+
+    return measurements
+  }
+
   @Test
   fun `reporting set is created and then retrieved`() = runBlocking {
     val measurementConsumerData = inProcessCmmsComponents.getMeasurementConsumerData()
@@ -537,7 +556,7 @@ abstract class InProcessLifeOfAReportIntegrationTest(
 
     assertThat(actualResult).reachValue().isWithin(tolerance).of(expectedResult.reach.value)
 
-    val measurements: List<Measurement> = inProcessCmmsComponents.listMeasurements()
+    val measurements: List<Measurement> = listMeasurements()
     assertThat(measurements).hasSize(1)
     assertThat(measurements[0].protocolConfig.protocolsList).hasSize(1)
     assertThat(measurements[0].protocolConfig.protocolsList[0].hasReachOnlyLiquidLegionsV2())
@@ -657,7 +676,7 @@ abstract class InProcessLifeOfAReportIntegrationTest(
 
     assertThat(actualResult).reachValue().isWithin(tolerance).of(expectedResult.reach.value)
 
-    val measurements: List<Measurement> = inProcessCmmsComponents.listMeasurements()
+    val measurements: List<Measurement> = listMeasurements()
     assertThat(measurements).hasSize(1)
     assertThat(measurements[0].protocolConfig.protocolsList).hasSize(1)
     assertThat(measurements[0].protocolConfig.protocolsList[0].hasHonestMajorityShareShuffle())
